+++ src/scorer/feature/chaos_scorer.ts
@@ -14,8 +14,11 @@ export default class ChaosScorer extends FeatureScorer {
     }
 
     async _score(toot: Toot) {
-        const existingScore = toot.getIndividualScore("raw", this.name);
-        if (existingScore) return existingScore;  // Return the existing score if it exists
+        // Return the existing score if it exists
+        if (toot.scoreInfo?.scores) {
+            const existingScore = toot.getIndividualScore("raw", this.name);
+            if (existingScore) return existingScore;
+        }
 
         try {
             return this.decimalHash(toot.reblog?.content || toot.content);